Refactor waste measurement summary into component for reuse
