+++ 2024-frontend/src/components/MeasurementDetail.vue
