+++ 2024-frontend/src/components/SourceChart.vue
